fix(queue): 경기별 승급 트랜잭션 분리 및 QueueService 정리
